added database

+++ src/main/java/server/games/Wheel.java
@@ -5,10 +5,10 @@ import server.ClientData;
 import java.util.Random;
 
 public class Wheel extends GameInstance {
-    private double userCoins;
+    private int userCoins;
     private double[] winRates = {0, 0, 0.5, 0.75,  1, 1, 1.25, 1.5,  1.5, 2};
 
-    public Wheel(double buyIn) {
+    public Wheel(int buyIn) {
         super(1, 1);
         this.userCoins = buyIn;
     }
@@ -36,7 +36,7 @@ public class Wheel extends GameInstance {
             Random generator = new Random();
             int randomIndex = generator.nextInt(winRates.length);
             System.out.println("Your ratio was: " + winRates[randomIndex]);
-            clientData.setCoins(clientData.getCoins()+userCoins*winRates[randomIndex]);
+            clientData.setCoins((int) ((double) (clientData.getCoins()+userCoins)*winRates[randomIndex]));
 
     }
